fix specialNode location
